+++ back/src/main/java/com/bezkoder/springjwt/payload/response/BodyPartResponse.java
@@ -3,14 +3,14 @@ package com.bezkoder.springjwt.payload.response;
 import java.util.List;
 
 public class BodyPartResponse {
-    private Integer id;
+    private Long id;
     private String name;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -25,7 +25,7 @@ public class BodyPartResponse {
         this.name = name;
     }
 
-    public BodyPartResponse(Integer id, String name) {
+    public BodyPartResponse(Long id, String name) {
         this.id = id;
         this.name = name;
     }
