+++ back/src/main/java/com/bezkoder/springjwt/repository/BodyPartRepository.java
@@ -3,5 +3,5 @@ package com.bezkoder.springjwt.repository;
 import com.bezkoder.springjwt.models.BodyPart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BodyPartRepository extends JpaRepository<BodyPart, Integer> {
+public interface BodyPartRepository extends JpaRepository<BodyPart, Long> {
 }
